fix littlesis spec
previously running `littlesis spec` would always print this error message:
"Inappropriate ioctl for device"

While it still started a usable shell, it was clear that not everywhere was
function normally because it had some strange quirks such
as the 'exit' command not actually exiting. Now there are no more error
messages and everything appears to work as it should.